+++ src/cli.py
@@ -1,6 +1,7 @@
 #!/usr/bin/python3
 
 from api import Interface
+from translate import tr
 
 CURSOR_UP_ONE = '\x1b[1A'
 ERASE_LINE = '\x1b[2K'
@@ -12,21 +13,21 @@ def printMessage(text, message=False, username=None):
     global interface
     if username:
         interface.otherUsername = username
-        return printMessage("Correspondant username changed to " + username)
+        return printMessage(tr("username.other.changed") + username)
     if message:
         username = interface.otherUsername
         if(not username):
-            username = "L'autre"
-        print("\r" + ERASE_LINE + username + " : " + text + "\nCommand : ", end="")
+            username = tr("user.other")
+        print("\r" + ERASE_LINE + username + " : " + text + "\n" + tr("command") + " : ", end="")
     else:
-        print("\r" + ERASE_LINE + text + "\nCommand : ", end="")
+        print("\r" + ERASE_LINE + text + "\n" + tr("command") + " : ", end="")
 
 
 interface = Interface(printMessage)
 
 while True:
     try:
-        text = input("\rCommand : ")
+        text = input("\r" + tr("command") + " : ")
     except KeyboardInterrupt:
         text = "/quit"
     print(CURSOR_UP_ONE + ERASE_LINE, end="")